requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
pip==22.0.3
pytz==2021.3
setuptools==60.8.2
sqlparse==0.4.2
wheel==0.37.1
